+++ src/schemathesis/filters.py
@@ -1,0 +1,23 @@
+import re
+from typing import Optional
+
+from .types import Filter
+from .utils import force_tuple
+
+
+def should_skip_method(method: str, pattern: Optional[Filter]) -> bool:
+    if pattern is None:
+        return False
+    patterns = force_tuple(pattern)
+    return method.upper() not in map(str.upper, patterns)
+
+
+def should_skip_endpoint(endpoint: str, pattern: Optional[Filter]) -> bool:
+    if pattern is None:
+        return False
+    patterns = force_tuple(pattern)
+    return not any(is_match(endpoint, item) for item in patterns)
+
+
+def is_match(endpoint: str, pattern: str) -> bool:
+    return pattern in endpoint or bool(re.search(pattern, endpoint))