+++ docs/changelog.rst
@@ -6,6 +6,9 @@ Changelog
 `Unreleased`_
 -------------
 
+`0.5.0`_ - 2019-09-16
+---------------------
+
 Added
 ~~~~~
 
@@ -82,7 +85,8 @@ Fixed
 
 - Initial public release
 
-.. _Unreleased: https://github.com/kiwicom/schemathesis/compare/v0.4.1...HEAD
+.. _Unreleased: https://github.com/kiwicom/schemathesis/compare/v0.5.0...HEAD
+.. _0.5.0: https://github.com/kiwicom/schemathesis/compare/v0.4.0...v0.5.0
 .. _0.4.1: https://github.com/kiwicom/schemathesis/compare/v0.4.0...v0.4.1
 .. _0.4.0: https://github.com/kiwicom/schemathesis/compare/v0.3.0...v0.4.0
 .. _0.3.0: https://github.com/kiwicom/schemathesis/compare/v0.2.0...v0.3.0